machineroom-main/requirements.txt
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
cryptography==36.0.1
docopt==0.6.2
docutils==0.18.1
idna==3.3
Kivy-Garden==0.1.4
pipreqs==0.4.11
pycparser==2.21
Pygments==2.10.0
PyMySQL==1.0.2
requests==2.26.0
urllib3==1.26.7
yarg==0.1.9
